+++ yellowstone-grpc-kafka/src/lib.rs
@@ -1,0 +1,9 @@
+#![deny(clippy::clone_on_ref_ptr)]
+#![deny(clippy::missing_const_for_fn)]
+#![deny(clippy::trivially_copy_pass_by_ref)]
+
+pub mod config;
+pub mod dedup;
+pub mod grpc;
+pub mod prom;
+pub mod version;